Change council screen to not show stats
The stats are still shown so that their hover text is there, but with a
transparent font so that they aren't actually visible
Removed the number from the hover text so now it just says competent,
poor, masterful, etc.
Moved the attribute icons into the space where the stat was so that it
doesn't look awkwardly empty.